Using floats for replays properly and bugfix
